refactor(mobile): split notify pipeline
- extract pipeline helpers
- add context helper tests
